requirements.txt
# Core web framework
streamlit>=1.31.0

# Database
sqlalchemy>=2.0.0

# Data science and ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
plotly>=5.18.0

# AI/LLM
google-generativeai>=0.3.0

# Environment management
python-dotenv>=1.0.0
